♻️ refactor(spa): move DesktopLayoutContainer out of the route tree
AGENTS.md keeps route trees to thin page segments and puts layout chunks under
`src/features/`, but this unit sat in `src/routes/(main)/_layout/` while two
things outside the tree reached into it: `ChatInput/Desktop` for
`LayoutContainerContext` (long-standing), and now `BootShell/AppShellSkeleton`
for the container's css variables and emotion classes, which it deliberately
shares so the boot skeleton and the real container cannot drift apart.

Sharing that unit is right; its address was wrong. The component, its context,
css-variable helpers and styles move together to
`src/features/DesktopLayoutContainer/`, and the four call sites now import from
`@/features/*`. No behavior change — git records four of the five files as
identical renames, and the fifth differs only by its three relative imports.

After this, nothing outside `src/routes/**` imports from `(main)/_layout`, and
`src/spa/BootShell` no longer depends on the route tree at all.

Note the wider debt this does not touch: several features still reach into other
route internals (`ResourceManager` and `Home` into
`@/routes/(main)/resource/features/*` and `@/routes/(main)/home/_layout/*`,
`useInterceptingRoutes` into the agent profile). That belongs in its own change.

Reported by Codex on #17829.
